Counting zeros in a number using recursion.

+++ 04-Recursion/count_zeros.cpp
@@ -2,7 +2,7 @@
 int count_zeros(const long &);
 
 int main() {
-    int number {0};
+    long number {0};
 
     std::cout << "Enter a number >> ";
     std::cin >> number;
